@@ -5,15 +5,23 @@
     "dependencies": {
         "@emotion/core": "^10.0.27",
         "@emotion/react": "^11.4.1",
-        "@emotion/styled": "^10.0.27",
+        "@emotion/styled": "^11.3.0",
         "@ibm/plex": "^5.1.0",
+        "@material-ui/core": "^4.12.3",
+        "@material-ui/icons": "^4.11.2",
+        "@material-ui/styles": "^4.11.4",
+        "@mui/core": "^5.0.0-alpha.47",
+        "@mui/icons-material": "^5.0.0",
+        "@mui/material": "^5.0.0",
+        "@mui/styles": "^5.0.0",
         "@projectstorm/react-canvas-core": "^6.6.1",
         "@projectstorm/react-diagrams": "^6.6.1",
         "@projectstorm/react-diagrams-core": "^6.6.1",
         "@projectstorm/react-diagrams-defaults": "^6.6.1",
         "@projectstorm/react-diagrams-routing": "^6.6.1",
         "@reduxjs/toolkit": "^1.6.1",
-        "@rjsf/core": "^3.0.2",
+        "@rjsf/core": "^3.1.0",
+        "@rjsf/material-ui": "^3.1.0",
         "@testing-library/jest-dom": "^4.2.4",
         "@testing-library/react": "^9.3.2",
         "@testing-library/user-event": "^7.1.2",
@@ -22,13 +30,11 @@
         "@types/lodash": "^4.14.150",
         "@types/node": "^16.6.1",
         "@types/react": "^17.0.18",
-        "@types/react-bootstrap": "^1.0.0",
         "@types/react-dom": "^17.0.9",
         "@types/react-redux": "^7.1.18",
         "await-spawn": "^4.0.2",
         "bitcoin-core": "^3.0.0",
         "bitcoinjs-lib": "^5.1.6",
-        "bootstrap": "^4.4.1",
         "closest": "^0.0.1",
         "color": "^3.1.2",
         "dagre": "^0.8.5",
@@ -39,9 +45,8 @@
         "pathfinding": "^0.4.18",
         "paths-js": "^0.4.10",
         "prettier": "^2.2.0",
-        "react": "^16.12.0",
-        "react-bootstrap": "^1.0.0",
-        "react-dom": "^16.12.0",
+        "react": "17.0.2",
+        "react-dom": "17.0.2",
         "react-redux": "^7.2.4",
         "react-scripts": "^4.0.3",
         "redux-thunk": "^2.3.0",
