@@ -1,277 +1,112 @@
+import { InputLabel, MenuItem, Select } from '@mui/material';
 import * as Bitcoin from 'bitcoinjs-lib';
 import React from 'react';
-import './InputDetail.css';
-import { OutpointDetail } from './OutpointDetail';
-import Form from 'react-bootstrap/Form';
+import { useDispatch } from 'react-redux';
 import {
-    hash_to_hex,
     sequence_convert,
     time_to_pretty_string,
+    txid_buf_to_string,
 } from '../../../util';
-import Hex from './Hex';
+import Hex, { ReadOnly } from './Hex';
+import './InputDetail.css';
+import { RefOutpointDetail } from './OutpointDetail';
 interface IProps {
     txinput: Bitcoin.TxInput;
     witnesses: Buffer[][];
-    psbts: Bitcoin.Psbt[];
-    goto: () => void;
-}
-interface IState {
-    open: boolean;
-    witness_selection: number | undefined;
-    psbt: Bitcoin.Psbt | undefined;
-    flash: null | any;
 }
 function maybeDecode(to_asm: boolean, elt: Buffer): string {
     if (to_asm) {
         return Bitcoin.script.toASM(
-            Bitcoin.script.decompile(elt) ?? new Buffer('')
+            Bitcoin.script.decompile(elt) ?? Buffer.from('')
         );
     } else {
         return elt.toString('hex');
     }
 }
-export class InputDetail extends React.Component<IProps, IState> {
-    form: any;
-    constructor(props: IProps) {
-        super(props);
-        this.state = {
-            open: false,
-            witness_selection: undefined,
-            psbt: undefined,
-            flash: null,
-        };
-        this.form = null;
-    }
-    async save_psbt(psbt: string) {
-        // no await
-        window.electron.save_psbt(psbt);
-    }
-    flash(msg: string | JSX.Element, color: string, onclick?: () => void) {
-        const click = onclick ?? (() => null);
-        this.setState({
-            flash: (
-                <h3 style={{ color: color }} onClick={click}>
-                    {msg}
-                </h3>
-            ),
-        });
-        setTimeout(() => this.setState({ flash: <div></div> }), 2000);
-    }
-    async combine_psbt(psbt: Bitcoin.Psbt) {
-        let psbt_in = Bitcoin.Psbt.fromBase64(
-            await window.electron.fetch_psbt()
-        );
-        try {
-            psbt.combine(psbt_in);
-            this.flash('PSBT Combined', 'green');
-        } catch (e: any) {
-            this.flash(
-                <div>
-                    PSBT Error{' '}
-                    <span className="glyphicon glyphicon-question-sign"></span>
-                </div>,
-                'red',
-                () => alert(e.toString())
-            );
-        }
-    }
-    async sign_psbt(psbt: string) {
-        const command = [{ method: 'walletprocesspsbt', parameters: [psbt] }];
-        const signed = (await window.electron.bitcoin_command(command))[0];
-        const as_psbt = Bitcoin.Psbt.fromBase64(signed.psbt);
-        if (signed.complete) {
-            this.flash('Fully Signed', 'green');
-        } else {
-            this.flash('Partial Signed', 'red');
-        }
-        return as_psbt;
-    }
-    async finalize_psbt(psbt: string) {
-        const command = [{ method: 'finalizepsbt', parameters: [psbt] }];
-        const result = (await window.electron.bitcoin_command(command))[0];
-        if (!result.complete || !result.hex) {
-            this.flash('PSBT Not Complete', 'red');
-            return;
-        }
-        try {
-            const hex_tx = Bitcoin.Transaction.fromHex(result.hex);
-            const send = [
-                { method: 'sendrawtransaction', parameters: [result.hex] },
-            ];
-
-            const sent = (await window.electron.bitcoin_command(send))[0];
-            if (sent !== hex_tx.getId()) {
-                this.flash(
-                    <div>
-                        Relay Error{' '}
-                        <span className="glyphicon glyphicon-question-sign"></span>
-                    </div>,
-                    'red',
-                    () => alert(sent.message.toString())
-                );
-            } else {
-                this.flash('Transaction Relayed!', 'green');
-            }
-        } catch (e: any) {
-            this.flash(
-                <div>
-                    PSBT Error{' '}
-                    <span className="glyphicon glyphicon-question-sign"></span>
-                </div>,
-                'red',
-                () => alert(e.toString())
-            );
-        }
-    }
-    render() {
-        let witness_display = null;
-        if (
-            this.state.witness_selection !== undefined &&
-            this.props.witnesses.length > this.state.witness_selection
-        ) {
-            const index: number = this.state.witness_selection;
-            witness_display = this.props.witnesses[index]!.map((elt, i) => (
-                <Hex key={i} className="txhex" value={maybeDecode(true, elt)} />
-            ));
-        }
-        const psbts_display =
-            this.state.psbt === undefined ? (
-                <>
-                    <div></div>
-                    <div></div>
-                </>
-            ) : (
-                <>
-                    <div className="PSBTActions">
-                        <div title="Save PSBT to Disk">
-                            <i
-                                className="glyphicon glyphicon-floppy-save SavePSBT"
-                                onClick={(() =>
-                                    this.save_psbt(
-                                        this.state.psbt!.toBase64()
-                                    )).bind(this)}
-                            ></i>
-                        </div>
-                        <div title="Sign PSBT Using Node Wallet">
-                            <i
-                                className="glyphicon glyphicon-pencil SignPSBT"
-                                onClick={(async () => {
-                                    const psbt = await this.sign_psbt(
-                                        this.state.psbt!.toBase64()
-                                    );
-                                    // TODO: Confirm this saves to model?
-                                    this.state.psbt?.combine(psbt);
-                                    this.setState({ psbt: this.state.psbt });
-                                }).bind(this)}
-                            ></i>
-                        </div>
-                        <div title="Combine PSBT from File">
-                            <i
-                                className="glyphicon glyphicon-compressed CombinePSBT"
-                                onClick={(async () => {
-                                    // TODO: Confirm this saves to model?
-                                    const psbt = await this.combine_psbt(
-                                        this.state.psbt!
-                                    );
-                                    this.setState({ psbt: this.state.psbt });
-                                }).bind(this)}
-                            ></i>
-                        </div>
-                        <div title="Attempt Finalizing and Broadcast">
-                            <i
-                                className="glyphicon glyphicon-send BroadcastPSBT"
-                                onClick={async () => {
-                                    await this.finalize_psbt(
-                                        this.state.psbt!.toBase64()
-                                    );
-                                }}
-                            ></i>
-                        </div>
-                    </div>
-                    <Hex
-                        className="txhex"
-                        value={this.state.psbt.toBase64()}
-                    ></Hex>
-                </>
-            );
-        const scriptValue = Bitcoin.script.toASM(
-            Bitcoin.script.decompile(this.props.txinput.script) ??
-                Buffer.from('Error Decompiling')
-        );
-        const seq = this.props.txinput.sequence;
-        const { relative_time, relative_height } = sequence_convert(seq);
-        const sequence =
-            relative_time === 0 ? (
-                relative_height === 0 ? null : (
-                    <div className="InputDetailSequence">
-                        <span>Relative Height: </span>
-                        {relative_height}
-                    </div>
-                )
-            ) : (
-                <div className="InputDetailSequence">
-                    <span>Relative Time: </span>
-                    {time_to_pretty_string(relative_time)}
-                </div>
-            );
-
-        const witness_options = this.props.witnesses.map((w, i) => (
-            <option key={i} value={i}>
-                {i}
-            </option>
+export function InputDetail(props: IProps) {
+    const dispatch = useDispatch();
+    const witness_selection_form = React.useRef<HTMLSelectElement>(null);
+    const [witness_selection, setWitness] = React.useState(0);
+    let witness_display = null;
+    if (
+        witness_selection !== undefined &&
+        props.witnesses.length > witness_selection
+    ) {
+        const index: number = witness_selection;
+        witness_display = props.witnesses[index]!.map((elt, i) => (
+            <Hex key={i} className="txhex" value={maybeDecode(true, elt)} />
         ));
-        const scriptSig =
-            this.props.txinput.script.length === 0 ? null : (
-                <div className="InputDetailScriptSig">
-                    <p>ScriptSig:</p>
-                    <Hex className="txhex" value={scriptValue}></Hex>
+    }
+    const scriptValue = Bitcoin.script.toASM(
+        Bitcoin.script.decompile(props.txinput.script) ??
+            Buffer.from('Error Decompiling')
+    );
+    const seq = props.txinput.sequence;
+    const { relative_time, relative_height } = sequence_convert(seq);
+    const sequence =
+        relative_time === 0 ? (
+            relative_height === 0 ? null : (
+                <div className="InputDetailSequence">
+                    <ReadOnly
+                        value={relative_height.toString()}
+                        label="Relative Height"
+                    ></ReadOnly>
                 </div>
-            );
-        // missing horizontal
-        return (
-            <div>
-                <OutpointDetail
-                    txid={hash_to_hex(this.props.txinput.hash)}
-                    n={this.props.txinput.index}
-                    onClick={() => this.props.goto()}
-                />
-                {sequence}
-                {scriptSig}
-                <Form
+            )
+        ) : (
+            <div className="InputDetailSequence">
+                <ReadOnly
+                    value={time_to_pretty_string(relative_time)}
+                    label="Relative Time"
+                ></ReadOnly>
+            </div>
+        );
+
+    const witness_options = props.witnesses.map((w, i) => (
+        <MenuItem key={i} value={i}>
+            {i}
+        </MenuItem>
+    ));
+    const scriptSig =
+        props.txinput.script.length === 0 ? null : (
+            <div className="InputDetailScriptSig">
+                <Hex
+                    className="txhex"
+                    value={scriptValue}
+                    label="ScriptSig"
+                ></Hex>
+            </div>
+        );
+    // missing horizontal
+    return (
+        <div className="InputDetail">
+            <RefOutpointDetail
+                txid={txid_buf_to_string(props.txinput.hash)}
+                n={props.txinput.index}
+            />
+            {sequence}
+            {scriptSig}
+            <form>
+                <InputLabel id="label-select-witness">Witness</InputLabel>
+                <Select
+                    labelId="label-select-witness"
+                    label="Witness"
+                    variant="outlined"
+                    ref={witness_selection_form}
                     onChange={() => {
-                        console.log(this.form.value);
-                        this.setState({
-                            witness_selection: this.form.value || undefined,
-                        });
-                        if (this.form.value) {
-                            this.setState({
-                                psbt: this.props.psbts[this.form.value],
-                            });
+                        const idx: number =
+                            parseInt(
+                                witness_selection_form.current?.value ?? '0'
+                            ) ?? 0;
+                        if (idx < props.witnesses.length && idx >= 0) {
+                            setWitness(idx);
                         }
                     }}
                 >
-                    <Form.Group>
-                        <Form.Label>
-                            <div> Witness: {this.state.witness_selection}</div>
-                        </Form.Label>
-                        <Form.Control
-                            as="select"
-                            ref={(r: any) => {
-                                this.form = r;
-                            }}
-                        >
-                            <option value={undefined}></option>
-                            {witness_options}
-                        </Form.Control>
-                    </Form.Group>
-                </Form>
-                {witness_display}
-                {this.state.flash}
-                <div className="InputDetailPSBT">
-                    <div> PSBT: </div>
-                    {psbts_display}
-                </div>
-            </div>
-        );
-    }
+                    {witness_options}
+                </Select>
+            </form>
+            <div>{witness_display}</div>
+        </div>
+    );
 }
