@@ -2,27 +2,19 @@ import {
     DefaultLinkModel,
     DefaultLinkModelOptions,
 } from '@projectstorm/react-diagrams';
+import { TransactionModel } from '../../../../Data/Transaction';
+import { UTXOModel } from '../../../../Data/UTXO';
 export class SpendLinkModel extends DefaultLinkModel {
-    is_reachable: boolean;
-    reachable_callback: (_: boolean) => void;
+    link_type: 'exclusive' | 'nonexclusive';
+    linked_to: TransactionModel | null;
+
     constructor(options?: DefaultLinkModelOptions) {
         super({
             type: 'spend',
             width: 15,
-            color: 'white',
             ...options,
         });
-        this.is_reachable = true;
-        this.reachable_callback = (b) => null;
-    }
-    setReachable(b: boolean) {
-        this.is_reachable = b;
-        this.reachable_callback(b);
-    }
-    isReachable(): boolean {
-        return this.is_reachable;
-    }
-    registerReachableCallback(c: (_: boolean) => void) {
-        this.reachable_callback = c;
+        this.link_type = 'exclusive';
+        this.linked_to = null;
     }
 }
