@@ -1,18 +1,12 @@
 import React from 'react';
-import Tooltip from 'react-bootstrap/Tooltip';
-import OverlayTrigger from 'react-bootstrap/OverlayTrigger';
-function BaseHex(props: { value: string; styling: string }) {
+import { Tooltip, TextField } from '@mui/material';
+import { useTheme } from '@mui/material';
+function BaseHex(props: { value: string; styling: string; label?: string }) {
     const [tip_message, set_tip] = React.useState(null as null | string);
-    const code = React.useRef(null as null | HTMLElement);
+    const code = React.useRef<HTMLDivElement>(null);
+    const theme = useTheme();
     const copy = () => {
-        const select = window.getSelection();
-        if (!select || !code.current) return;
-        select.removeAllRanges();
-        const range = document.createRange();
-        range.selectNodeContents(code.current);
-        select.addRange(range);
-        const txt = select.toString().trim();
-        navigator.clipboard.writeText(txt);
+        navigator.clipboard.writeText(props.value);
         set_tip('Copied!');
         setTimeout(() => {
             set_tip(null);
@@ -23,24 +17,52 @@ function BaseHex(props: { value: string; styling: string }) {
         code.current?.addEventListener('dblclick', copy);
     });
     return (
-        <OverlayTrigger
+        <Tooltip
+            title={tip_message ?? 'double click to copy.'}
+            arrow
             placement="top"
-            overlay={
-                <Tooltip id={'hex-copyable-' + Math.random().toString()}>
-                    {tip_message ?? 'double click to copy.'}
-                </Tooltip>
-            }
         >
-            <code className={props.styling} ref={code}>
-                {props.value}{' '}
-            </code>
-        </OverlayTrigger>
+            <div>
+                <TextField
+                    ref={code}
+                    fullWidth
+                    label={props.label}
+                    defaultValue={props.value}
+                    variant="outlined"
+                    InputProps={{
+                        readOnly: true,
+                    }}
+                />
+            </div>
+        </Tooltip>
     );
 }
-export default function Hex(props: { value: string; className?: string }) {
+export default function Hex(props: {
+    value: string;
+    className?: string;
+    label?: string;
+}) {
     return BaseHex({ ...props, styling: 'truncate ' + props.className });
 }
 
-export function ASM(props: { value: string; className?: string }) {
+export function ASM(props: {
+    value: string;
+    className?: string;
+    label?: string;
+}) {
     return BaseHex({ ...props, styling: 'ASM ' + props.className });
+}
+
+export function ReadOnly(props: { value: React.ReactNode; label: string }) {
+    return (
+        <TextField
+            fullWidth
+            label={props.label}
+            defaultValue={props.value}
+            variant="outlined"
+            InputProps={{
+                readOnly: true,
+            }}
+        />
+    );
 }
