@@ -1,45 +1,64 @@
+import { IconButton, Tooltip } from '@mui/material';
+import { green } from '@mui/material/colors';
+import DoubleArrowIcon from '@mui/icons-material/DoubleArrow';
 import React from 'react';
-import Button from 'react-bootstrap/esm/Button';
-import ListGroup from 'react-bootstrap/ListGroup';
+import { useDispatch } from 'react-redux';
+import { select_txn, select_utxo } from '../EntitySlice';
 import Hex from './Hex';
 import './OutpointDetail.css';
-export class OutpointDetail extends React.Component<{
-    txid: string;
-    n: number;
-    onClick: () => void;
-}> {
-    render() {
-        return (
-            <div>
-                <span>Outpoint:</span>
-                <div className="OutpointDetail">
-                    <Hex
-                        className="txhex"
-                        value={this.props.txid.toString() + ':' + this.props.n}
-                    />
+export function OutpointDetail(props: { txid: string; n: number }) {
+    const dispatch = useDispatch();
+    return (
+        <div className="OutpointDetail">
+            <Hex
+                className="txhex"
+                label="Outpoint"
+                value={props.txid.toString() + ':' + props.n}
+            />
+            <Tooltip title="Go To the Transaction that created this.">
+                <IconButton
+                    aria-label="goto-creating-txn"
+                    onClick={() => dispatch(select_txn(props.txid))}
+                >
+                    <DoubleArrowIcon style={{ color: green[500] }} />
+                </IconButton>
+            </Tooltip>
+        </div>
+    );
+}
 
-                    <Button variant="link" onClick={() => this.props.onClick()}>
-                        <span
-                            className="glyphicon glyphicon-chevron-right"
-                            style={{ color: 'green' }}
-                            title="Go to the transaction that created this."
-                        ></span>
-                    </Button>
-                </div>
-            </div>
-        );
-    }
+export function RefOutpointDetail(props: { txid: string; n: number }) {
+    const dispatch = useDispatch();
+    return (
+        <div className="OutpointDetail">
+            <Hex
+                className="txhex"
+                label="Outpoint"
+                value={props.txid.toString() + ':' + props.n}
+            />
+            <Tooltip title="Go to this outpoint">
+                <IconButton
+                    aria-label="goto-this-outpoint"
+                    onClick={() =>
+                        dispatch(
+                            select_utxo({
+                                hash: props.txid,
+                                nIn: props.n,
+                            })
+                        )
+                    }
+                >
+                    <DoubleArrowIcon style={{ color: green[500] }} />
+                </IconButton>
+            </Tooltip>
+        </div>
+    );
 }
 
 export class TXIDDetail extends React.Component<{
     txid: string;
 }> {
     render() {
-        return (
-            <div className="TXIDDetail">
-                <span>txid:</span>
-                <Hex className="txhex" value={this.props.txid} />
-            </div>
-        );
+        return <Hex className="txhex" value={this.props.txid} label="TXID" />;
     }
 }
