@@ -1,19 +1,22 @@
+import { IconButton, Tooltip, useTheme } from '@mui/material';
+import { red } from '@mui/material/colors';
+import CancelOutlinedIcon from '@mui/icons-material/CancelOutlined';
 import React from 'react';
-import { ContractModel, Data } from '../../Data/ContractManager';
-import { TransactionDetail } from './Detail/TransactionDetail';
-import { UTXODetail } from './Detail/UTXODetail';
+import { useDispatch, useSelector } from 'react-redux';
+import { ContractModel } from '../../Data/ContractManager';
 import { TransactionModel } from '../../Data/Transaction';
 import { UTXOModel } from '../../Data/UTXO';
-import './EntityViewer.css';
-import Button from 'react-bootstrap/esm/Button';
-import { OutpointInterface, TXID, TXIDAndWTXIDMap } from '../../util';
-import { useDispatch, useSelector } from 'react-redux';
+import { TXIDAndWTXIDMap } from '../../util';
+import { TransactionDetail } from './Detail/TransactionDetail';
+import { UTXODetail } from './Detail/UTXODetail';
 import {
     deselect_entity,
     EntityType,
     selectEntityToView,
     selectShouldViewEntity,
 } from './EntitySlice';
+import './EntityViewer.css';
+import Color from 'color';
 export interface ViewableEntityInterface {}
 
 interface CurrentylViewedEntityProps {
@@ -21,6 +24,7 @@ interface CurrentylViewedEntityProps {
 }
 
 export function CurrentlyViewedEntity(props: CurrentylViewedEntityProps) {
+    const theme = useTheme();
     const [width, setWidth] = React.useState('20em');
     const onMouseUp = (e: MouseEvent) => {
         e.preventDefault();
@@ -83,18 +87,24 @@ export function CurrentlyViewedEntity(props: CurrentylViewedEntityProps) {
     }
     const dispatch = useDispatch();
     return (
-        <div className="EntityViewerFrame">
+        <div
+            className="EntityViewerFrame"
+            style={{
+                backgroundColor: Color(theme.palette.background.default)
+                    .fade(0.2)
+                    .toString(),
+            }}
+        >
             <div className="EntityViewerResize" onMouseDown={onMouseDown}></div>
             <div>
-                <Button
-                    onClick={() => dispatch(deselect_entity())}
-                    variant="link"
-                >
-                    <span
-                        className="glyphicon glyphicon-remove"
-                        style={{ color: 'red' }}
-                    ></span>
-                </Button>
+                <Tooltip title="Close Entity Viewer">
+                    <IconButton
+                        aria-label="close-entity-viewer"
+                        onClick={() => dispatch(deselect_entity())}
+                    >
+                        <CancelOutlinedIcon style={{ color: red[500] }} />
+                    </IconButton>
+                </Tooltip>
                 <div
                     className="EntityViewer"
                     style={{
